+++ src/code/core/profiler.py
@@ -5,7 +5,6 @@ import json
 import logging
 import datetime
 
-from core.helpers import get_utc_now
 from core.config import settings as cfg
 
 
@@ -37,30 +36,6 @@ class Profiler:
     def get(self, name, default_value=None):
         self._get()
         return self.data.get(name, default_value)
-    
-    def get_records_remaining(self):
-        if cfg.rate_limit_period <= 0 and cfg.rate_limit_records <= 0:
-            return cfg.records_per_request
-        
-        timestamp = get_utc_now() - datetime.timedelta(minutes=cfg.rate_limit_period)
-        history_data = self.get("history", [])
-        r_count = 0
-        i = len(history_data) - 1
-        while i >= 0:
-            item = history_data[i]
-            ts = datetime.datetime.strptime(str(item[0]), "%Y-%m-%d %H:%M:%S")
-            if ts > timestamp:
-                r_count = r_count + item[1]
-            else:
-                del history_data[i]
-
-            i = i - 1
-
-        # Put it back for next time
-        self.set("history", history_data)
-
-        r_count = cfg.rate_limit_records - r_count
-        return r_count if r_count > 0 else 0
 
     def update(self):
         self._get()
@@ -101,23 +76,6 @@ class Profiler:
                 pass
             else:
                 return
-
-        if name == "history":
-            if isinstance(value, list):
-                for x in value:
-                    if not isinstance(x, list):
-                        return
-                    
-                    if len(x) != 2:
-                        return
-                    
-                    if isinstance(x[1], int):
-                        return
-                    
-                    try:
-                        s = datetime.datetime.strptime(str(x[0]), "%Y-%m-%d %H:%M:%S")
-                    except:
-                        return
 
         self.data[name] = value
 
@@ -175,7 +133,7 @@ class LocalProfiler(Profiler):
             return False
 
         return True
-    
+
 
 class DynamoDBProfiler(Profiler):
     def __init__(self, **kwargs):